--- module/io/__index__.ts
@@ -1,1 +1,0 @@
-export * from './src/util';
