--- module/io/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@travetto/io",
-  "version": "5.0.0-rc.10",
-  "description": "IO Utilities for travetto applications.",
-  "keywords": [
-    "io",
-    "travetto",
-    "typescript"
-  ],
-  "homepage": "https://travetto.io",
-  "license": "MIT",
-  "author": {
-    "email": "travetto.framework@gmail.com",
-    "name": "Travetto Framework"
-  },
-  "files": [
-    "__index__.ts",
-    "src",
-    "support"
-  ],
-  "main": "__index__.ts",
-  "repository": {
-    "url": "https://github.com/travetto/travetto.git",
-    "directory": "module/io"
-  },
-  "engines": {
-    "node": ">=22.0.0"
-  },
-  "dependencies": {
-    "@travetto/runtime": "^5.0.0-rc.5",
-    "file-type": "^16.5.4",
-    "@types/mime": "^3.0.4",
-    "mime": "^3.0.0"
-  },
-  "travetto": {
-    "displayName": "IO Support"
-  },
-  "private": false,
-  "publishConfig": {
-    "access": "public"
-  }
-}
