--- module/io/support/fixtures/asset.yml
@@ -1,3 +1,0 @@
-asset:
-  yaml:
-    value: 10
