+++ module/model/test/blob.ts
@@ -9,20 +9,21 @@ class BlobUtilSuite {
   @Test()
   async simpleTest() {
     const meta: BlobMeta = {
-      hash: 'ora_nge_bee_for_sly_'
+      hash: 'ora_nge_bee_for_sly_',
+      filename: 'bob'
     };
 
     assert(ModelBlobUtil.getHashedLocation(meta) === 'ora_/nge_/bee_/for_/sly_.bin');
 
-    meta.contentType = 'image/jpeg';
+    meta.filename = 'billy.jpeg';
 
     assert(ModelBlobUtil.getHashedLocation(meta) === 'ora_/nge_/bee_/for_/sly_.jpeg');
 
-    meta.contentType = 'video/mp4';
+    meta.filename = 'video.mp4';
 
     assert(ModelBlobUtil.getHashedLocation(meta) === 'ora_/nge_/bee_/for_/sly_.mp4');
 
-    meta.contentType = 'application/octet-stream';
+    meta.filename = 'none';
 
     assert(ModelBlobUtil.getHashedLocation(meta) === 'ora_/nge_/bee_/for_/sly_.bin');
   }
@@ -36,6 +37,7 @@ class BlobUtilSuite {
     assert(ModelBlobUtil.getHashedLocation(meta) === 'ora_/nge_/bee.bin');
 
     meta.contentType = 'image/jpeg';
+    meta.filename = 'image.jpeg';
 
     assert(ModelBlobUtil.getHashedLocation(meta) === 'ora_/nge_/bee.jpeg');
