+++ module/rest-upload/support/fixtures/asset.yml
@@ -1,0 +1,2 @@
+---
+yml: text
