+++ module/rest-upload/support/fixtures/logo.gif
