+++ module/rest-upload/support/fixtures/logo.png
