+++ module/rest-upload/support/test/server.ts
@@ -1,11 +1,10 @@
 import assert from 'node:assert';
 
-import { BlobUtil } from '@travetto/runtime';
+import { BlobUtil, BytesUtil } from '@travetto/runtime';
 import { Controller, Post, Request } from '@travetto/rest';
 import { BaseRestSuite } from '@travetto/rest/support/test/base';
 import { BeforeAll, Suite, Test, TestFixtures } from '@travetto/test';
 import { RootRegistry } from '@travetto/registry';
-import { IOUtil } from '@travetto/io';
 
 import { Upload, UploadAll } from '../../src/decorator';
 
@@ -69,7 +68,7 @@ export abstract class RestUploadServerSuite extends BaseRestSuite {
     const res = await this.request<{ hash: string }>('post', '/test/upload/all', this.getMultipartRequest([sent]));
 
     const file = await this.fixture.readStream('/logo.png');
-    assert(res.body.hash === await IOUtil.hashInput(file));
+    assert(res.body.hash === await BytesUtil.hashInput(file));
   }
 
 
@@ -85,7 +84,7 @@ export abstract class RestUploadServerSuite extends BaseRestSuite {
     });
 
     const file = await this.fixture.readStream('/logo.png');
-    assert(res.body.hash === await IOUtil.hashInput(file));
+    assert(res.body.hash === await BytesUtil.hashInput(file));
   }
 
   @Test()
@@ -94,7 +93,7 @@ export abstract class RestUploadServerSuite extends BaseRestSuite {
     const res = await this.request<{ hash: string }>('post', '/test/upload', this.getMultipartRequest(uploads));
 
     const file = await this.fixture.readStream('/logo.png');
-    assert(res.body.hash === await IOUtil.hashInput(file));
+    assert(res.body.hash === await BytesUtil.hashInput(file));
   }
 
   @Test()
@@ -105,7 +104,7 @@ export abstract class RestUploadServerSuite extends BaseRestSuite {
     );
     const res = await this.request<{ hash1: string, hash2: string }>('post', '/test/upload/all-named', this.getMultipartRequest(uploads));
     const file = await this.fixture.readStream('/logo.png');
-    const hash = await IOUtil.hashInput(file);
+    const hash = await BytesUtil.hashInput(file);
 
     assert(res.body.hash1 === hash);
     assert(res.body.hash2 === hash);
@@ -136,10 +135,10 @@ export abstract class RestUploadServerSuite extends BaseRestSuite {
     assert(res.status === 200);
 
     const file1 = await this.fixture.readStream('/logo.gif');
-    const hash1 = await IOUtil.hashInput(file1);
+    const hash1 = await BytesUtil.hashInput(file1);
 
     const file2 = await this.fixture.readStream('/logo.png');
-    const hash2 = await IOUtil.hashInput(file2);
+    const hash2 = await BytesUtil.hashInput(file2);
 
     assert(res.body.hash1 === hash1);
     assert(res.body.hash2 === hash2);
@@ -170,10 +169,10 @@ export abstract class RestUploadServerSuite extends BaseRestSuite {
     assert(res.status === 200);
 
     const file1 = await this.fixture.readStream('/asset.yml');
-    const hash1 = await IOUtil.hashInput(file1);
+    const hash1 = await BytesUtil.hashInput(file1);
 
     const file2 = await this.fixture.readStream('/logo.png');
-    const hash2 = await IOUtil.hashInput(file2);
+    const hash2 = await BytesUtil.hashInput(file2);
 
     assert(res.body.hash1 === hash1);
     assert(res.body.hash2 === hash2);
